requirements.txt
opencv-python==4.1.0.25
Cython==0.29.5
datetime==4.3
Flask==1.1.1
Flask-Cors==3.0.7
numpy==1.17.2
numpydoc==0.8.0
pandas==0.24.1
Pillow==6.2.0
tqdm==4.31.1
unicodecsv==0.14.1
urllib3==1.24.1
utils==0.9.0
virtualenv==16.7.5
Werkzeug==0.16.0
gunicorn==19.9.0
